refactor(builder): extract builder-types.ts (R1)
