+++ packages/llm-agent-libs/src/builder.ts
@@ -33,7 +33,6 @@ import type {
   ISubpromptClassifier,
   IToolCache,
   IToolSelectionStrategy,
-  SmartAgentHandle as SmartAgentHandleBase,
   SmartAgentRagStores,
   SubAgentRegistry,
 } from '@mcp-abap-adt/llm-agent';
@@ -59,6 +58,11 @@ import {
 import { makeConnectionStrategy } from '@mcp-abap-adt/llm-agent-mcp';
 import { wrapEmbedder } from './adapters/usage-logging-embedder.js';
 import { SmartAgent, type SmartAgentConfig } from './agent.js';
+import type {
+  SmartAgentBuilderConfig,
+  SmartAgentHandle,
+} from './builder-types.js';
+import { isModelProvider } from './builder-types.js';
 import {
   LlmClassifier,
   type LlmClassifierConfig,
@@ -87,11 +91,7 @@ import { DefaultPipeline } from './pipeline/default-pipeline.js';
 import type { DagCoordinatorHandlerDeps } from './pipeline/handlers/dag-coordinator.js';
 import type { IStageHandler } from './pipeline/stage-handler.js';
 import type { IPluginLoader } from './plugins/types.js';
-import type {
-  IPromptInjectionDetector,
-  IToolPolicy,
-  SessionPolicy,
-} from './policy/types.js';
+import type { IPromptInjectionDetector, IToolPolicy } from './policy/types.js';
 import type { IReranker } from './reranker/types.js';
 import { RateLimiterLlm } from './resilience/rate-limiter-llm.js';
 import { RetryLlm } from './resilience/retry-llm.js';
@@ -104,81 +104,18 @@ import { SmartAgentSubAgent } from './subagent/smart-agent-subagent.js';
 import type { ITracer } from './tracer/types.js';
 import type { IOutputValidator } from './validator/types.js';
 
-// ---------------------------------------------------------------------------
-// Config types
-// ---------------------------------------------------------------------------
-
-export interface BuilderMcpConfig {
-  type: 'http' | 'stdio';
-  /** HTTP: MCP endpoint URL */
-  url?: string;
-  /** stdio: command to spawn */
-  command?: string;
-  /** stdio: command arguments */
-  args?: string[];
-  /** HTTP headers (e.g. x-sap-destination for reverse proxy routing) */
-  headers?: Record<string, string>;
-}
-
-export interface BuilderPromptsConfig {
-  /** Preamble prepended to the ContextAssembler system message. */
-  system?: string;
-  /** Override the intent-classifier system prompt. */
-  classifier?: string;
-  /** Instruction for the reasoning/strategy block. */
-  reasoning?: string;
-  /** Prompt for query translation for RAG. */
-  ragTranslate?: string;
-  /** Prompt for history summarization. */
-  historySummary?: string;
-}
-
-export interface SmartAgentBuilderConfig {
-  /** MCP connection(s). Pass an array to connect multiple servers simultaneously. */
-  mcp?: BuilderMcpConfig | BuilderMcpConfig[];
-  /** SmartAgent orchestration limits. */
-  agent?: Partial<SmartAgentConfig>;
-  /** System / classifier prompt overrides. */
-  prompts?: BuilderPromptsConfig;
-  /** Data governance policy for RAG records. */
-  sessionPolicy?: SessionPolicy;
-  /** Skip startup model validation (useful for testing). Default: false. */
-  skipModelValidation?: boolean;
-  /** Attempts for the startup model-validation chat before aborting — lenient
-   *  retry on ANY transient failure (e.g. a SAP AI Core deployment-list blip at
-   *  boot). Default 3. */
-  modelValidationAttempts?: number;
-  /** Base backoff between startup-validation attempts (× attempt number).
-   *  Default 2000ms. */
-  modelValidationBackoffMs?: number;
-}
-
-// ---------------------------------------------------------------------------
-// Handle returned by build()
-// ---------------------------------------------------------------------------
-
-/**
- * SmartAgentHandle specialized for the concrete SmartAgent class.
- *
- * This re-exports the generic SmartAgentHandle from @mcp-abap-adt/llm-agent
- * with SmartAgent as the type parameter, preserving full concrete typing
- * (including internal methods like `applyConfigUpdate`, `reconfigure`,
- * `getActiveConfig`) for callers in llm-agent-libs and llm-agent-server.
- */
-export type SmartAgentHandle = SmartAgentHandleBase<SmartAgent>;
+// Re-export the public builder config/handle types so the package barrel
+// (index.ts → builder.js) and all external importers stay byte-stable.
+export type {
+  BuilderMcpConfig,
+  BuilderPromptsConfig,
+  SmartAgentBuilderConfig,
+  SmartAgentHandle,
+} from './builder-types.js';
 
 // ---------------------------------------------------------------------------
 // SmartAgentBuilder
 // ---------------------------------------------------------------------------
-
-function isModelProvider(obj: unknown): obj is IModelProvider {
-  return (
-    obj !== null &&
-    typeof obj === 'object' &&
-    typeof (obj as IModelProvider).getModels === 'function' &&
-    typeof (obj as IModelProvider).getModel === 'function'
-  );
-}
 
 export class SmartAgentBuilder {
   private readonly cfg: SmartAgentBuilderConfig;
